pyproject.toml
[tool.poetry]
name = "gedhtml"
version = "0.1.0"
description = "Converting GED files to static web pages."
authors = ["Edo van Veen <edo@vanveen.io>"]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
python-gedcom = "^1.0.0"
matplotlib = "^3.8.4"
yattag = "^1.15.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.1"
pytest = "^8.1.1"
mypy = "^1.10.0"

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.lint]
ignore-init-module-imports = false

[tool.mypy]
exclude = ["tests"]
ignore_missing_imports = true

[build-system]
requires = ["poetry_core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
